direktive-addon/consumer_script/requirements.txt
paho-mqtt>=2.0.0
opencv-python-headless
requests 